requirements.txt
pandas
Pillow
datetime
Flask
itsdangerous
Jinja2
MarkupSafe
Click
Werkzeug
gunicorn
requests
